refactor: adding helpers js
