+++ lib/functions.js
@@ -1,13 +1,3 @@
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
-
-import { createRequire } from 'module';
-
-const require = createRequire(import.meta.url);
-const { dirname: _dirname, join } = path;
-
-const dirname = _dirname(fileURLToPath(import.meta.url));
 
 export const getYYYYMMDDHHMMSS = (date = new Date()) => {
   return [
@@ -29,50 +19,3 @@ export const shuffleArray = (array) => {
   }
   return array;
 };
-
-export const readAsset = (assetPath) => {
-  return fs.readFileSync(join(dirname, './assets', assetPath)).toString();
-};
-
-export const pathConfig = async (options) => {
-  let PWD = process.env.PWD ? process.env.PWD : process.cwd();
-  if (options.pwdPath)
-    PWD = options.pwdPath.slice(0, 2) === './' ? join(PWD, options.pwdPath) : options.pwdPath;
-
-  let sequelizercPath = join(PWD, '.sequelizerc');
-  if (options.sequelizercPath) sequelizercPath = join(PWD, options.sequelizercPath);
-
-  // eslint-disable-next-line import/no-dynamic-require
-  const sequelizerc = (fs.existsSync(sequelizercPath)) ? require(sequelizercPath) : [];
-
-  const packageDir = join(PWD, 'package.json');
-
-  // @ToDo Add env option
-  
-  let migrationsDir = join(PWD, 'migrations');
-  let modelsDir = join(PWD, 'models');
-  let stateDir = modelsDir;
-  let indexDir;
-
-  if (sequelizerc['models-path'] && !options.ignoreSequelizerc)
-    modelsDir = sequelizerc['models-path'];
-  if (sequelizerc['migrations-path'] && !options.ignoreSequelizerc)
-    migrationsDir = sequelizerc['migrations-path'];
-
-  indexDir = join(modelsDir, 'index.js');
-
-  if (options.modelsPath) modelsDir = join(PWD, options.modelsPath);
-  if (options.migrationsPath) migrationsDir = join(PWD, options.migrationsPath);
-  if (options.indexFilePath) indexDir = join(PWD, options.indexFilePath);
-  if (options.statePath) stateDir = join(PWD, options.statePath);
-
-  return {
-    modelsDir,
-    migrationsDir,
-    stateDir,
-    indexDir,
-    packageDir,
-  };
-};
-
-export default pathConfig;
