+++ lib/migration.js
@@ -4,7 +4,8 @@ import lodash from 'lodash';
 
 import prettier from 'prettier';
 
-import { readAsset, getYYYYMMDDHHMMSS } from './functions.js';
+import { getYYYYMMDDHHMMSS } from './functions.js';
+import { readAsset } from './helpers.js';
 
 const { template } = lodash;
