+++ lib/models.js
@@ -188,90 +188,3 @@ export const parseDifference = (previousState, currentState) => {
   });
   return actions;
 };
-
-export const sortActions = (actions) => {
-  const orderedActionTypes = [
-    'removeIndex',
-    'removeColumn',
-    'dropTable',
-    'createTable',
-    'addColumn',
-    'changeColumn',
-    'addIndex',
-  ];
-
-  // test
-  // actions = shuffleArray(actions);
-
-  actions.sort((a, b) => {
-    if (orderedActionTypes.indexOf(a.actionType) < orderedActionTypes.indexOf(b.actionType))
-      return -1;
-    if (orderedActionTypes.indexOf(a.actionType) > orderedActionTypes.indexOf(b.actionType))
-      return 1;
-
-    if (a.depends.length === 0 && b.depends.length > 0) return -1; // a < b
-    if (b.depends.length === 0 && a.depends.length > 0) return 1; // b < a
-
-    return 0;
-  });
-
-  // #region new method for sorting actions / NEEDS More TESTING
-  /*
-  actions.sort((a, b) => {
-    if (
-      a.depends?.length !== 0 &&
-      b.depends?.length !== 0 &&
-      a.actionType === b.actionType &&
-      b.depends?.indexOf(a.tableName) !== -1
-    )
-      return -1;
-    return 0;
-  });
-
-  actions.forEach(() => {
-    actions.forEach((a, i) => {
-      if (a.depends?.length !== 0) {
-        actions.forEach((b, j) => {
-          if (
-            b.depends?.length !== 0 &&
-            a.actionType === b.actionType &&
-            b.depends?.indexOf(a.tableName) !== -1 &&
-            i > j
-          ) {
-            const temp = actions[i];
-            actions[i] = actions[j];
-            actions[j] = temp;
-          }
-        });
-      }
-    });
-  });
- */
-  // #endregion
-
-  for (let k = 0; k <= actions.length; k++) {
-    for (let i = 0; i < actions.length; i++) {
-      const a = actions[i];
-      if (a.depends?.length !== 0) {
-        for (let j = 0; j < actions.length; j++) {
-          const b = actions[j];
-          if (
-            b.depends?.length !== 0 &&
-            a.actionType === b.actionType &&
-            b.depends?.indexOf(a.tableName) !== -1 &&
-            i > j
-          ) {
-            const c = actions[i];
-            actions[i] = actions[j];
-            actions[j] = c;
-          }
-        }
-      }
-    }
-  }
-
-  // remove duplicate changeColumns
-  for (let i = 0; i < actions.length; i++) {
-    if (JSON.stringify(actions[i]) === JSON.stringify(actions[i - 1])) actions.splice(i, 1);
-  }
-};
