+++ lib/parser.js
@@ -1,113 +1,8 @@
 import Sequelize from 'sequelize';
 import hash from 'object-hash';
+import { getDefaultValueType, getColumnTypeName } from './helpers.js';
 
 const { log } = console;
-const { DataTypes } = Sequelize;
-
-export const getColumnTypeNameBase = (attrName, attrObj, options) => {
-  switch (attrName) {
-    // CHAR(length, binary)
-    case DataTypes.CHAR.key:
-      if (options.binary) return `CHAR.BINARY`;
-      return `CHAR(${options.length})`;
-
-    // STRING(length, binary).BINARY
-    case DataTypes.STRING.key:
-      return `STRING${options.length ? `(${options.length})` : ''}${
-        options.binary ? '.BINARY' : ''
-      }`;
-
-    // TEXT(length)
-    case DataTypes.TEXT.key:
-      if (!options.length) return `TEXT`;
-      if (typeof options.length === 'string') return `TEXT("${options.length.toLowerCase()}")`;
-      return `TEXT(${options.length.toLowerCase()})`;
-
-    // NUMBER(length, decimals).UNSIGNED.ZEROFILL
-    case DataTypes.NUMBER.key:
-    case DataTypes.TINYINT.key:
-    case DataTypes.SMALLINT.key:
-    case DataTypes.MEDIUMINT.key:
-    case DataTypes.BIGINT.key:
-    case DataTypes.FLOAT.key:
-    case DataTypes.REAL.key:
-    case DataTypes.DOUBLE.key:
-    case DataTypes.DECIMAL.key:
-    case DataTypes.INTEGER.key: {
-      let ret = attrName;
-      if (options.length) {
-        ret += `(${options.length}`;
-        if (options.decimals) ret += `, ${options.decimals}`;
-        ret += ')';
-      }
-
-      if (options.precision) {
-        ret += `(${options.precision}`;
-        if (options.scale) ret += `, ${options.scale}`;
-        ret += ')';
-      }
-
-      ret = [ret];
-
-      if (options.zerofill) ret.push('ZEROFILL');
-      if (options.unsigned) ret.push('UNSIGNED');
-
-      return `${ret.join('.')}`;
-    }
-
-    case DataTypes.ENUM.key:
-      return `ENUM('${options.values.join("', '")}')`;
-
-    case DataTypes.BLOB.key:
-      if (!options.length) return `BLOB`;
-      if (typeof options.length === 'string') return `BLOB("${options.length.toLowerCase()}")`;
-      return `BLOB(${options.length.toLowerCase()})`;
-
-    case DataTypes.GEOMETRY.key:
-      if (options.type) {
-        if (options.srid) return `GEOMETRY('${options.type}', ${options.srid})`;
-        return `GEOMETRY('${options.type}')`;
-      }
-      return 'GEOMETRY';
-
-    case DataTypes.GEOGRAPHY.key:
-      return 'GEOGRAPHY';
-
-    case DataTypes.ARRAY.key: {
-      const type = attrObj.toString();
-      const match = type.match(/(\w+)\[\]/);
-      const arrayType = match.length >= 1 ? `{prefix}${match[1]}` : '{prefix}STRING';
-      return `ARRAY(${arrayType})`;
-    }
-    case DataTypes.RANGE.key:
-      console.warn(`${attrName} type not supported, you should make it by`);
-      return attrObj.toSql();
-
-    // BOOLEAN, TIME, DATE, DATEONLY, HSTORE, JSONB
-    // UUID, UUIDV1, UUIDV4, VIRTUAL, INET, MACADDR
-    default:
-      return attrName;
-  }
-};
-
-export const getColumnTypeName = (col, prefix = 'Sequelize.') => {
-  const attrName = typeof col.type.key === 'undefined' ? col.type : col.type.key;
-  const attrObj = col.type;
-  const options = col.type.options ? col.type.options : {};
-
-  const columnTypeName = getColumnTypeNameBase(attrName, attrObj, options);
-
-  return `${prefix}${columnTypeName.replace('{prefix}', prefix)}`;
-};
-
-export const getDefaultValueType = (defaultValue, prefix = 'Sequelize.') => {
-  if (typeof defaultValue === 'object' && defaultValue.constructor && defaultValue.constructor.name)
-    return { internal: true, value: prefix + defaultValue.constructor.name };
-
-  if (typeof defaultValue === 'function') return { notSupported: true, value: '' };
-
-  return { value: defaultValue };
-};
 
 export const parseIndex = (idx) => {
   delete idx.parser;
@@ -137,7 +32,7 @@ export const parseIndex = (idx) => {
   return idx;
 };
 
-export const parseProperties = (modelName, columnName, column) => {
+export const parseProperties = (modelName, column) => {
   Object.entries(column).forEach(([propertyName, property]) => {
     if (
       propertyName.startsWith('_') ||
@@ -150,7 +45,7 @@ export const parseProperties = (modelName, columnName, column) => {
       if (!val.notSupported) {
         column[propertyName] = val;
       } else {
-        log(`[Not supported] Skip defaultValue column of attribute ${modelName}:${columnName}`);
+        log(`[Not supported] Skip defaultValue column of attribute ${modelName}:${column.field}`);
         delete column[propertyName];
       }
     }
@@ -162,7 +57,7 @@ export const parseAttributes = (sequelize, modelName, attributes) => {
     delete column.Model;
     delete column.fieldName;
     // delete attributes[attributeName].field;
-    parseProperties(modelName, columnName, column);
+    parseProperties(modelName, column);
 
     if (typeof column.type === 'undefined') {
       if (!column.seqType) {
